Now it looks like a game
The rounds system is now working properly. Player objects are always deleted normally (there are no more situations when there are 5 corpses running around the scene). At the end of the game the winner is marked in firebase and money is added, with the loser similarly. Returned back sliders with bars xp local player and opponent. Display the total score of rounds. Display of individual score of each player. There are still a couple of bugs with PhotonView, when the wrong link goes to the one somehow. Bug with an error in the master client in the lobby has not been fixed yet.